cmd/agent: add Linux build constraint
Node info on non-linux systems will default to zero.
